requirements.txt
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.5.0
ansi2html==1.8.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.4
attrs==23.1.0
autocommand==2.2.2
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
certipy==0.1.3
cffi==1.15.1
charset-normalizer==3.2.0
cheroot==10.0.0
CherryPy==18.8.0
click==8.1.7
comm==0.1.4
cryptography==41.0.3
d6tcollect==1.0.6
d6tflow==0.2.6
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
docutils==0.20.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fanstatic==1.3
fastjsonschema==2.18.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-resources==6.0.1
inflect==7.0.0
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.collections==4.3.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
jedi==0.19.0
Jinja2==3.1.2
js==1.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-telemetry==0.1.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterhub==4.0.2
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
lockfile==0.12.2
luigi==3.3.0
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==10.1.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pamela==1.1.0
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.16.1
plotly-express==0.4.1
portend==3.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyOpenSSL==23.2.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywebio==1.8.2
pywebio-battery==0.5.0
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
resend==0.5.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scipy==1.11.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
speaklater==1.3
SQLAlchemy==2.0.20
stack-data==0.6.2
statsmodels==0.14.0
tempora==5.5.0
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
uri-template==1.3.0
urllib3==2.0.4
user-agents==2.2.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.6.2
Werkzeug==2.2.3
widgetsnbextension==4.0.8
WTForms==3.0.1
zc.lockfile==3.0.post1
